requirements.txt
pandas
numpy
matplotlib
seaborn
xgboost
scikit-learn
scipy
python-dotenv
joblib
pydantic
fastapi
uvicorn